pyproject.toml
# ImageTrim 多平台打包配置

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ImageTrim"
version = "1.0.0"
description = "现代化的图片去重和格式转换工具"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "BlueRainCoat", email = "blueraincoatli@example.com"},
]
keywords = ["image", "deduplication", "avif", "converter", "pyqt6"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Utilities",
]
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.9.1",
    "Pillow>=11.3.0",
    "pillow-avif-plugin>=1.5.2",
    "requests>=2.32.5",
    "ImageHash>=4.3.2",
    "numpy>=2.3.3",
    "scipy>=1.16.2",
    "PyWavelets>=1.9.0",
    "send2trash>=1.8.3",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "setuptools>=61.0",
    "wheel",
]

[project.urls]
Homepage = "https://github.com/blueraincoatli/DeDupImg"
Repository = "https://github.com/blueraincoatli/DeDupImg"
Documentation = "https://github.com/blueraincoatli/DeDupImg/wiki"
"Bug Tracker" = "https://github.com/blueraincoatli/DeDupImg/issues"

[project.scripts]
imagetrim = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = [
    "resources/**/*",
    "resources/icons/**/*",
    "resources/images/**/*",
]

[tool.setuptools.data-files]
"share/icons/hicolor/256x256/apps" = ["app/resources/icons/imageTrim256px.png"]
"share/applications" = ["packaging/linux/imagetrim.desktop"]
"share/metainfo" = ["packaging/linux/imagetrim.metainfo.xml"]